* 'master' of https://github.com/samiroquai/Henallux20172018:
  fix d'une mauvaise conversion de type

+++ BD Avancees et applications web/Labo4Octobre/BankDAL/BankAccountManager.cs
@@ -121,7 +121,7 @@ namespace BankDAL
                 connection.Open();
                 var command = new SqlCommand("SELECT [Solde] FROM [CompteEnBanque] WHERE [IBAN]=@iban", connection);
                 command.Parameters.AddWithValue("@iban", iban);
-                var solde = (double)command.ExecuteScalar();
+                var solde = (decimal)command.ExecuteScalar();
                 connection.Close();
                 return Convert.ToDouble(solde);
             }
